moved 'src' folder content & deleted it
