+++ badges/goldBadge.js
