+++ badges/index.js
