+++ badges/platinumBadge.js
