+++ helpers/augurAPI.js
