+++ helpers/awardBadge.js
